Sync ACPICA with Intel's version 20201217.
* Fix exception code class checks.

* Better GCC 10 support.

* Some -Wimplicit-fallthrough fixes.

* Miscellaneous improvements in the tools.

For detailed list, please see sys/contrib/dev/acpica/changes.txt.

+++ sys/contrib/dev/acpica/source/components/debugger/dbinput.c
@@ -644,7 +644,7 @@ AcpiDbGetNextToken (
 
     /* Remove any spaces at the beginning, ignore blank lines */
 
-    while (*String && isspace (*String))
+    while (*String && isspace ((int) *String))
     {
         String++;
     }
@@ -756,7 +756,7 @@ AcpiDbGetNextToken (
 
         /* Find end of token */
 
-        while (*String && !isspace (*String))
+        while (*String && !isspace ((int) *String))
         {
             String++;
         }
